Handle the widening of FuncCallContext's call_cntr.
This one is a bit stickier, because the value is passed to Java code
that implements ResultSetProvider, an interface in pljava-api. The
signature for assignRowValues can't just be changed.

Instead, for now, a ResultSetProvider is limited to returning INT_MAX
rows, but the limit is checked to ensure a predictable failure.

For the future, in the Java 8 world, the ResultSetProvider interface
could grow a default method largeAssignRowValues that takes a long
row number; the Java 8 default implementation would just signal the
same error, and classes implementing the old interface won't break.
Alas, PL/Java isn't committed to supporting only Java 8 yet.

In fact, the same thing could be done on the quiet: PL/Java could check
whether the class implementing ResultSetProvider happens to have a
largeAssignRowValues method with the right signature, and use it if
present. It would simply have to be covered in documentation, and not
formally declared in the interface until the Java support bar moves to
Java 8.

All of that is future work - this change merely checks and enforces
an INT_MAX limit.

+++ pljava-so/src/main/c/type/Type.c
@@ -415,7 +415,8 @@ Datum Type_invokeSRF(Type self, jclass cls, jmethodID method, jvalue* args, PG_F
 	currentInvocation->hasConnected = ctxData->hasConnected;
 	currentInvocation->invocation   = ctxData->invocation;
 
-	hasRow = Type_hasNextSRF(self, ctxData->rowProducer, ctxData->rowCollector, (jint)context->call_cntr);
+	hasRow = Type_hasNextSRF(self, ctxData->rowProducer, ctxData->rowCollector,
+		(jlong)context->call_cntr);
 
 	ctxData->hasConnected = currentInvocation->hasConnected;
 	ctxData->invocation   = currentInvocation->invocation;
@@ -629,7 +630,7 @@ static jobject _Type_getSRFCollector(Type self, PG_FUNCTION_ARGS)
 	return 0;
 }
 
-static bool _Type_hasNextSRF(Type self, jobject rowProducer, jobject rowCollector, jint callCounter)
+static bool _Type_hasNextSRF(Type self, jobject rowProducer, jobject rowCollector, jlong callCounter)
 {
 	return (JNI_callBooleanMethod(rowProducer, s_Iterator_hasNext) == JNI_TRUE);
 }
@@ -656,7 +657,7 @@ jobject Type_getSRFCollector(Type self, PG_FUNCTION_ARGS)
 	return self->typeClass->getSRFCollector(self, fcinfo);
 }
 
-bool Type_hasNextSRF(Type self, jobject rowProducer, jobject rowCollector, jint callCounter)
+bool Type_hasNextSRF(Type self, jobject rowProducer, jobject rowCollector, jlong callCounter)
 {
 	return self->typeClass->hasNextSRF(self, rowProducer, rowCollector, callCounter);
 }
